Factory Design Patern
* Add DriverManagerInterface
* Add DriverManagerFactory
* Add FirefoxDriverManager
* Add ChromeDriverManager
