+++ master-selenium-framework/src/test/java/dev/ericmarcelo/selenium/pom/base/BaseJUnitTest.java
@@ -1,6 +1,6 @@
 package dev.ericmarcelo.selenium.pom.base;
 
-import dev.ericmarcelo.selenium.pom.factory.DriverManager;
+import dev.ericmarcelo.selenium.pom.factory.DriverManagerOriginal;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.openqa.selenium.WebDriver;
@@ -9,7 +9,7 @@ public class BaseJUnitTest {
     protected WebDriver webDriver;
     @BeforeEach
     public void startDriver() {
-      webDriver = new DriverManager().initializeDriver("FIREFOX");
+      webDriver = new DriverManagerOriginal().initializeDriver("FIREFOX");
 
     }
