+++ master-selenium-framework/src/test/java/dev/ericmarcelo/selenium/pom/factory/ChromeDriverManager.java
@@ -1,0 +1,10 @@
+package dev.ericmarcelo.selenium.pom.factory;
+
+import org.openqa.selenium.WebDriver;
+
+public class ChromeDriverManager implements DriverManager{
+    @Override
+    public WebDriver createDriver() {
+        return null;
+    }
+}
