+++ master-selenium-framework/src/test/java/dev/ericmarcelo/selenium/pom/factory/DriverManagerOriginal.java
@@ -1,0 +1,40 @@
+package dev.ericmarcelo.selenium.pom.factory;
+
+import dev.ericmarcelo.selenium.pom.constants.DriverType;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+import java.io.File;
+
+public class DriverManagerOriginal {
+
+    public WebDriver initializeDriver(String browser) {
+
+        WebDriver webDriver;
+
+        switch(DriverType.valueOf(browser)) {
+            case FIREFOX -> {
+                WebDriverManager
+                        .firefoxdriver()
+                        .cachePath("src"
+                                + File.separator
+                                + "test"
+                                + File.separator
+                                + "resources"
+                                + File.separator
+                                + "drivers")
+                        .setup();
+
+                webDriver = new FirefoxDriver();
+                webDriver.manage().window().maximize();
+                break;
+            }
+            default -> {
+                throw new IllegalArgumentException("Invalid browser name: " + browser);
+            }
+        }
+
+        return webDriver;
+    }
+}
