+++ master-selenium-framework/src/test/resources/drivers/resolution.properties
@@ -1,8 +1,8 @@
 #WebDriverManager Resolution Cache
-#Mon Mar 27 15:41:05 PDT 2023
+#Tue Mar 28 08:02:26 PDT 2023
 firefox110-ttl=14\:34\:51 23/03/2023 PDT
 firefox111-ttl=10\:08\:50 28/03/2023 PDT
 firefox110=0.32.2
 firefox=111
 firefox111=0.32.2
-firefox-ttl=16\:41\:05 27/03/2023 PDT
+firefox-ttl=09\:02\:26 28/03/2023 PDT
